Run testcafe tests in concurrency mode (#15332)

+++ testing/testcafe/runner.js
@@ -16,12 +16,18 @@ createTestCafe('localhost', 1437, 1438)
 
 
         componentFolder = componentFolder ? `${componentFolder}/**` : '**';
+        if(fs.existsSync('./testing/testcafe/screenshots')) {
         // eslint-disable-next-line spellcheck/spell-checker
-        fs.rmdirSync('./testing/testcafe/screenshots', { recursive: true });
+            fs.rmdirSync('./testing/testcafe/screenshots', { recursive: true });
+        }
         const runner = testCafe.createRunner()
             .browsers(args.browsers.split(' '))
+
             .src([`./testing/testcafe/tests/${componentFolder}/${file}.ts`]);
 
+        if(args.concurrency > 0) {
+            runner.concurrency(args.concurrency);
+        }
         if(testName) {
             runner.filter(name => name === testName);
         }
@@ -40,6 +46,7 @@ createTestCafe('localhost', 1437, 1438)
 function getArgs() {
     return parseArgs(process.argv.slice(1), {
         default: {
+            concurrency: 0,
             browsers: 'chrome',
             test: '',
             componentFolder: '',
